# postman_and_swagger_specs/user_service_swagger.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Finance Management Application",
        "description": "This service responsible for maintaining users",
        "termsOfService": "FMA T&C",
        "contact": {
            "name": "Yashwanth Kumar HT",
            "email": "ht.yashwanthkumar@gmail.com"
        },
        "license": {
            "name": "FMA Licence LLC"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:7070",
            "description": "Generated server url"
        }
    ],
    "paths": {
        "/user/{userId}": {
            "get": {
                "tags": [
                    "user-controller"
                ],
                "summary": "Get User Details",
                "operationId": "fetchUserDetails",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User Details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserDto"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "User not found"
                    },
                    "500": {
                        "description": "Service failed due to internal server error while fetching User details",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResponseBodyUserDto"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "user-controller"
                ],
                "summary": "Update User Details",
                "operationId": "updateUserDetails",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserDto"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "User details updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserDto"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Service failed due to internal server error while updating User details",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResponseBodyLong"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "user-controller"
                ],
                "summary": "Delete User",
                "operationId": "deleteUser",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserDto"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Service failed due to internal server error while deleting User details",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResponseBodyVoid"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/user": {
            "get": {
                "tags": [
                    "user-controller"
                ],
                "summary": "Get All Users Details",
                "operationId": "fetchAllUsersDetails",
                "responses": {
                    "200": {
                        "description": "All Users Details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserDto"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Service failed due to internal server error while fetching Users details",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResponseBodyListUserDto"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "user-controller"
                ],
                "summary": "Save User Details",
                "operationId": "saveUser",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserDto"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "User saved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserDto"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Service failed due to internal server error while saving User",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResponseBodyLong"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "UserDto": {
                "type": "object",
                "properties": {
                    "pkUserId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "title": {
                        "type": "string",
                        "minLength": 1
                    },
                    "firstName": {
                        "type": "string",
                        "minLength": 1
                    },
                    "lastName": {
                        "type": "string",
                        "minLength": 1
                    },
                    "email": {
                        "type": "string",
                        "minLength": 1
                    },
                    "phoneNumber": {
                        "type": "string",
                        "minLength": 1
                    }
                }
            },
            "ResponseBodyLong": {
                "type": "object",
                "properties": {
                    "message": {
                        "type": "string"
                    },
                    "data": {
                        "type": "integer",
                        "format": "int64"
                    }
                }
            },
            "ResponseBodyListUserDto": {
                "type": "object",
                "properties": {
                    "message": {
                        "type": "string"
                    },
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/UserDto"
                        }
                    }
                }
            },
            "ResponseBodyUserDto": {
                "type": "object",
                "properties": {
                    "message": {
                        "type": "string"
                    },
                    "data": {
                        "$ref": "#/components/schemas/UserDto"
                    }
                }
            },
            "ResponseBodyVoid": {
                "type": "object",
                "properties": {
                    "message": {
                        "type": "string"
                    },
                    "data": {
                        "type": "object"
                    }
                }
            }
        }
    }
}